Separated the unit tests after Sofiane's suggestion
