+++ python/test/t_GaussianLinearCalibration_std.py
@@ -65,7 +65,6 @@ def testCalibrationtExponential():
         # Analysis of the results
         # Maximum A Posteriori estimator
         thetaMAP = calibrationResult.getParameterMAP()
-        print("thetaMAP",thetaMAP)
         exactTheta = ot.Point([5.69186,0.0832132,0.992301])
         rtol = 1.e-2
         assert_almost_equal(thetaMAP, exactTheta, rtol)
@@ -91,7 +90,6 @@ def testCalibrationtExponential():
         # Analysis of the results
         # Maximum A Posteriori estimator
         thetaMAP = calibrationResult.getParameterMAP()
-        print("thetaMAP",thetaMAP)
         exactTheta = ot.Point([3.4397,0.095908,0.99096])
         rtol = 1.e-2
         assert_almost_equal(thetaMAP, exactTheta, rtol)
@@ -109,107 +107,5 @@ def testCalibrationtExponential():
         print("result=", calibrationResult)
     return
 
-def testCalibrationtChaboche():
-    '''
-    Thanks to - Antoine Dumas, Phimeca
-    References
-    - J. Lemaitre and J. L. Chaboche (2002) "Mechanics of solid materials" Cambridge University Press.
-    '''
-    ot.RandomGenerator.SetSeed(0)
-    
-    # Create the function.
-    g = ot.SymbolicFunction(["strain","R","C","gam"],["R + C*(1-exp(-gam*strain))"]) 
-    
-    # Define the random vector.
-    Strain = ot.Uniform(0,0.07)
-    unknownR = 750
-    unknownC = 2750
-    unknownGamma = 10
-    R = ot.Dirac(unknownR)
-    C = ot.Dirac(unknownC)
-    Gamma = ot.Dirac(unknownGamma)
-    
-    Strain.setDescription(["Strain"])
-    R.setDescription(["R"])
-    C.setDescription(["C"])
-    Gamma.setDescription(["Gamma"])
-    
-    # Create the joint input distribution function.
-    inputRandomVector = ot.ComposedDistribution([Strain, R, C, Gamma])
-    
-    # Create the Monte-Carlo sample.
-    sampleSize = 10
-    inputSample = inputRandomVector.getSample(sampleSize)
-    outputStress = g(inputSample)
-        
-    # Generate observation noise.
-    stressObservationNoiseSigma = 40. # (Pa)
-    noiseSigma = ot.Normal(0.,stressObservationNoiseSigma)
-    sampleNoiseH = noiseSigma.getSample(sampleSize)
-    observedStress = outputStress + sampleNoiseH
-    observedStrain = inputSample[:,0]
-    
-    # Set the calibration parameters
-    # Define the value of the reference values of the $\theta$ parameter. 
-    # In the bayesian framework, this is called the mean of the *prior* gaussian distribution. 
-    # In the data assimilation framework, this is called the *background*.
-    R = 700 # Exact : 750
-    C = 2500 # Exact : 2750
-    Gamma = 8. # Exact : 10
-    thetaPrior = ot.Point([R,C,Gamma])
-    
-    # The following statement create the calibrated function from the model. 
-    # The calibrated parameters Ks, Zv, Zm are at indices 1, 2, 3 in the inputs arguments of the model.
-    calibratedIndices = [1,2,3]
-    mycf = ot.ParametricFunction(g, calibratedIndices, thetaPrior)
-    
-    # Gaussian linear calibration
-    # The standard deviation of the observations.
-    sigmaStress = 10. # (Pa)
-
-    # Define the covariance matrix of the output Y of the model.
-    errorCovariance = ot.CovarianceMatrix(1)
-    errorCovariance[0,0] = sigmaStress**2
-    
-    # Define the covariance matrix of the parameters $\theta$ to calibrate.
-    sigmaR = 0.1 * R
-    sigmaC = 0.1 * C
-    sigmaGamma = 0.1 * Gamma
-    
-    sigma = ot.CovarianceMatrix(3)
-    sigma[0,0] = sigmaR**2
-    sigma[1,1] = sigmaC**2
-    sigma[2,2] = sigmaGamma**2
-    
-    # Calibration
-    # The `GaussianLinearCalibration` class performs the gaussian linear 
-    # calibration by linearizing the model in the neighbourhood of the prior.
-    algo = ot.GaussianLinearCalibration(mycf, observedStrain, observedStress, \
-                                        thetaPrior, sigma, errorCovariance)
-    
-    # The `run` method computes the solution of the problem.
-    algo.run()
-    
-    calibrationResult = algo.getResult()
-    
-    # Analysis of the results
-    # Maximum A Posteriori estimator
-    thetaMAP = calibrationResult.getParameterMAP()
-    print("Chaboche, thetaMAP",thetaMAP)
-    exactTheta = ot.Point([762.661,3056.59,8.52781])
-    assert_almost_equal(thetaMAP, exactTheta)
-    
-    # Covariance matrix of theta
-    thetaPosterior = calibrationResult.getParameterPosterior()
-    covarianceThetaStar = matrixToSample(thetaPosterior.getCovariance())
-    exactCovarianceTheta = ot.Sample(\
-        [[ 42.4899, 288.43, -1.70502 ],\
-         [ 288.43, 15977.1, -67.6046 ],\
-         [ -1.70502, -67.6046, 0.294659 ]])
-    assert_almost_equal(covarianceThetaStar, exactCovarianceTheta)
-    
-    return
-
 testCalibrationtExponential()
-testCalibrationtChaboche()
